requirements.txt
fastapi~=0.112.0
alembic~=1.13.2
sqlalchemy~=2.0.31
python-dotenv~=1.0.1
starlette~=0.37.2
greenlet
asyncpg
psycopg2-binary
uvicorn
itsdangerous
pydantic~=2.8.2
redis~=5.0.8
aioredis
aiogram~=3.10.0
celery~=5.4.0
requests~=2.32.3
aiohttp~=3.9.5
